fts_ts: retry i2c commands if it fails at first
The i2c devices we are connecting to may not be ready yet, or just
busy. Retry with a 10ms pause in order to still obtain touch data.
This improves overall touch data robustness, but the main thing it
does:
WE CAN FINALLY GET TOUCH DATA IN SUSPEND ALL THE TIME.

Signed-off-by: Dennis Rassmann <showp1984@gmail.com>
Signed-off-by: arter97 <qkrwngud825@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>